Volume Time

+++ app/src/main/java/com/shufflteam/shuffl/Point.java
@@ -1,23 +1,23 @@
 package com.shufflteam.shuffl;
 
 public class Point {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Point(int x, int y){
+    public Point(float x, float y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setTo(int x, int y){
+    public void setTo(float x, float y){
         this.x = x;
         this.y = y;
     }
